requirements.txt
requests
Pillow
labelme